+++ generators/@types/business-object-client-type.generator.ts
@@ -26,7 +26,11 @@ const businessObjectClientTypeGenerator: DesignGenerator = {
     {
       metadataType: 'BusinessObject',
       condition: (metadata) => !isLibrary(metadata),
-    }
+    },
+    {
+      metadataType: 'Behavior',
+      condition: (metadata) => !isLibrary(metadata),
+    },
   ],
 
   outputs: (metadata: DesignMetadata) => [