+++ generators/@types/business-object-type.generator.ts
@@ -27,7 +27,11 @@ const businessObjectTypeGenerator: DesignGenerator = {
     {
       metadataType: 'BusinessObject',
       condition: (metadata) => !isLibrary(metadata),
-    }
+    },
+    {
+      metadataType: 'Behavior',
+      condition: (metadata) => !isLibrary(metadata),
+    },
   ],
 
   outputs: (metadata: DesignMetadata) => [